Move ValveSchedulerTest to Catch2
