--- test-old/main.cpp
@@ -1,9 +1,0 @@
-#include <gtest/gtest.h>
-
-int main(int argc, char **argv)
-{
-    ::testing::InitGoogleTest(&argc, argv);
-    // if you plan to use GMock, replace the line above with
-    // ::testing::InitGoogleMock(&argc, argv);
-    return RUN_ALL_TESTS();
-}
